Windows fixes for VC++ compiler compability
myisam/myisam_ftdump.c:
  VC++ compiler compability fix
mysys/my_getsystime.c:
  Applied Sergei's code for Windows (still subject to changes by him)
sql/handler.cc:
  VC++ compiler compability fix
sql/item_geofunc.cc:
  Removed non-used variable
sql/item_strfunc.cc:
  VC++ compiler compability fix
sql/opt_range.cc:
  VC++ compiler compability fix
sql/sql_insert.cc:
  VC++ compiler compability fix
sql/sql_lex.cc:
  VC++ compiler compability fix
sql/sql_parse.cc:
  VC++ compiler compability fix
sql/sql_prepare.cc:
  VC++ compiler compability fix
sql/sql_union.cc:
  Removed non-used variable and VC++ compiler compability fix

+++ sql/item_strfunc.cc
@@ -2825,9 +2825,9 @@ String *Item_func_uuid::val_str(String *str)
   uuid_time=tv;
   pthread_mutex_unlock(&LOCK_uuid_generator);
 
-  uint32 time_low=            tv & 0xFFFFFFFF;
-  uint16 time_mid=            (tv >> 32) & 0xFFFF;
-  uint16 time_hi_and_version= (tv >> 48) | UUID_VERSION;
+  uint32 time_low=            (uint32) (tv & 0xFFFFFFFF);
+  uint16 time_mid=            (uint16) ((tv >> 32) & 0xFFFF);
+  uint16 time_hi_and_version= (uint16) ((tv >> 48) | UUID_VERSION);
 
   str->realloc(UUID_LENGTH+1);
   str->length(UUID_LENGTH);
